// src/app/interfaces/assessment-report.interface.ts
export class IAssessmentReport {
    grade: number[];
    program: number;
    programSopi: string;
    sopi:string;
    programCourse: number;
    course:string;
    assessmentLevel: number;
    assessmentTask: string;
    passingGrade: number;
    target: number;
    assessmentComment: string;
    performance: number;
    improvementPlan: string;
    assessmentCycle: number;
    term: number;
    academicYear: string;
    id:number;
}